delete confirmation

+++ src/SearchResults.js
@@ -116,7 +116,12 @@ class SearchResults extends Component {
       <button
                       className="bg-blueGray-800 text-white active:bg-blueGray-600 text-sm font-bold uppercase px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 w-full ease-linear transition-all duration-150"
                       type="button"  onClick={(e) =>{ e.preventDefault();
-                        this.deleteFlight(fl.FlightNumber);
+                        if (window.confirm("Are you sure you want to delete the flight?")) {
+                         this.deleteFlight(fl.FlightNumber);
+
+                      } else {
+                        
+                        }
                         
                             }
                     }
